fixing cg

+++ tellez.aron/P3/functions.cpp
@@ -20,7 +20,7 @@ char* remplaceChar(const char* str, size_t size, char replaceFrom, char replaceT
   return result;
 }
 
-int hasRepeatedChars(const char* str, size_t size)
+size_t hasRepeatedChars(const char* str, size_t size)
 {
   for (size_t i = 0; i < size; ++i)
   {
