+++ tellez.aron/P3/functions.hpp
@@ -2,5 +2,5 @@
 #define FUNCTIONS_HPP
 #include <cstddef>
 char* remplaceChar(const char* str, size_t size, char replaceFrom, char replaceTo);
-int hasRepeatedChars(const char* str, size_t size);
+size_t hasRepeatedChars(const char* str, size_t size);
 #endif
